move exposure to components
