+++ cbsurge/components/builtenv/__init__.py
@@ -1,5 +1,5 @@
 import click
-from cbsurge.exposure.builtenv.buildings import buildings
+from cbsurge.components.builtenv.buildings import buildings
 @click.group()
 def builtenv():
     f"""Command line interface for {__package__} package"""
