+++ cbsurge/components/builtenv/buildings/fgb.py
@@ -2,9 +2,9 @@
 import random
 import threading
 from collections import deque
-from cbsurge.exposure.builtenv.buildings.fgbgdal import OVERPASS_API_URL, GMOSM_BUILDINGS_ROOT
-from pyogrio.raw import open_arrow, write_arrow, read
-from cbsurge.exposure.builtenv.buildings.pmt import WEB_MERCATOR_TMS
+from cbsurge.components.builtenv.buildings.fgbgdal import OVERPASS_API_URL, GMOSM_BUILDINGS_ROOT
+from pyogrio.raw import open_arrow
+from cbsurge.components.builtenv.buildings.pmt import WEB_MERCATOR_TMS
 from cbsurge.constants import ARROWTYPE2OGRTYPE
 import morecantile as m
 from cbsurge import util
@@ -19,7 +19,7 @@ from pyproj import Geod
 import math
 import concurrent
 from pyarrow import compute as pc
-from rich.progress import Progress, TaskProgressColumn,BarColumn,SpinnerColumn, TimeRemainingColumn, TextColumn
+from rich.progress import Progress, TaskProgressColumn,BarColumn, TimeRemainingColumn, TextColumn
 import httpx
 from osm2geojson import json2geojson
 
@@ -470,7 +470,6 @@ def download_admin(admin_path=None, out_path=None, country_col_name=None, admin_
 
 
 if __name__ == '__main__':
-    import asyncio
     logger = util.setup_logger(name='rapida', level=logging.INFO)
 
     nf = 5829
